Relocate diet goals files
